requirements.txt
Flask
request
python-dotenv